Add pay api swift template
- Add pay api swift template
- Add versioning
