+++ README.md
@@ -137,8 +137,6 @@ modelProtocol | customize protocol name that all models conform to | `String` | 
 enumNames | override enum names | `[String: String]` | [:]
 enumUndecodableCase | whether to add undecodable case to enums | `Bool` | false
 propertyNames | override property names | `[String: String]` | [:]
-safeArrayDecoding | filter out invalid items in array instead of throwing | `Bool` | false
-safeOptionalDecoding | set invalid optionals to nil instead of throwing | `Bool` | false
 tagPrefix | prefix for all tags | `String` | null
 tagSuffix | suffix for all tags | `String` | null
 codableResponses | constrains all responses to be `Codable` | `Bool` | false
