+++ Templates/Swift/Common/APIResult.swift
@@ -1,3 +1,3 @@
-{% include "Includes/Header.stencil" %}
+{% include "Common/Includes/Header.stencil" %}
 
 public typealias APIResult<T> = Result<T, APIClientError>
