+++ Templates/Swift/Common/Enum.swift
@@ -1,0 +1,3 @@
+{% include "Common/Includes/Header.stencil" %}
+
+{% include "Common/Includes/Enum.stencil" %}
