+++ Templates/Swift/Common/Includes/Enum.stencil
