+++ Templates/Swift/Common/Includes/Header.stencil
