+++ Templates/Swift/Common/Includes/Model.stencil
@@ -2,7 +2,7 @@
 /** {{ description }} */
 {% endif %}
 {% if enum %}
-{% include "Includes/Enum.stencil" enum %}
+{% include "Common/Includes/Enum.stencil" enum %}
 {% elif aliasType %}
 public typealias {{ type }} = {{ aliasType }}
 {% elif additionalPropertiesType and allProperties.count == 0 %}
@@ -52,7 +52,7 @@ public {{ options.modelType }} {{ type }}: {% if parent %}{{ parent.type }}{% el
     {% for enum in enums %}
     {% if not enum.isGlobal %}
 
-    {% filter indent:4 %}{% include "Includes/Enum.stencil" enum %}{% endfilter %}
+    {% filter indent:4 %}{% include "Common/Includes/Enum.stencil" enum %}{% endfilter %}
     {% endif %}
     {% endfor %}
     {% for property in properties %}
@@ -68,7 +68,7 @@ public {{ options.modelType }} {{ type }}: {% if parent %}{{ parent.type }}{% el
     {% endif %}
     {% for schema in schemas %}
 
-    {% filter indent:4 %}{% include "Includes/Model.stencil" schema %}{% endfilter %}
+    {% filter indent:4 %}{% include "Common/Includes/Model.stencil" schema %}{% endfilter %}
     {% endfor %}
 
     public {% if parent %}{% if properties.count == 0 %}override {% endif %}{% endif %}init({% for property in allProperties %}{{ property.name }}: {{ property.optionalType }}{% ifnot property.required %} = nil{% endif %}{% ifnot forloop.last %}, {% endif %}{% endfor %}) {
