+++ Templates/Swift/Common/Model.swift
@@ -1,0 +1,5 @@
+{% include "Common/Includes/Header.stencil" %}
+
+import Foundation
+
+{% include "Common/Includes/Model.stencil" %}
