+++ Templates/Swift/Common/URLEncoding.swift
@@ -7,8 +7,32 @@
 
 import Foundation
 
+enum APIRequestError: Error {
+    case encodingURL
+}
+
+extension CharacterSet { // Got it from Alamofire, because swift CharacterSet includes colons
+    /// Creates a CharacterSet from RFC 3986 allowed characters.
+    ///
+    /// RFC 3986 states that the following characters are "reserved" characters.
+    ///
+    /// - General Delimiters: ":", "#", "[", "]", "@", "?", "/"
+    /// - Sub-Delimiters: "!", "$", "&", "'", "(", ")", "*", "+", ",", ";", "="
+    ///
+    /// In RFC 3986 - Section 3.4, it states that the "?" and "/" characters should not be escaped to allow
+    /// query strings to include a URL. Therefore, all "reserved" characters with the exception of "?" and "/"
+    /// should be percent-escaped in the query string.
+    public static let apiURLQueryAllowed: CharacterSet = {
+        let generalDelimitersToEncode = ":#[]@" // does not include "?" or "/" due to RFC 3986 - Section 3.4
+        let subDelimitersToEncode = "!$&'()*+,;="
+        let encodableDelimiters = CharacterSet(charactersIn: "\(generalDelimitersToEncode)\(subDelimitersToEncode)")
+
+        return CharacterSet.urlQueryAllowed.subtracting(encodableDelimiters)
+    }()
+}
+
 struct URLEncoding {
-    static func encodeParams(_ params: [String: Any]) -> String {
+    static func encodeParams(_ params: [String: Any]) -> [(String, String)]  {
         var components: [(String, String)] = []
         for (key, value) in params {
             switch value {
@@ -34,7 +58,7 @@ struct URLEncoding {
             }
         }
 
-        return components.map { "\($0)=\($1)" }.joined(separator: "&")
+        return components
     }
 
     static func queryComponents(fromKey key: String, value: Any) -> [(String, String)] {
