--- Templates/Swift/Sources/Enum.swift
@@ -1,3 +1,0 @@
-{% include "Includes/Header.stencil" %}
-
-{% include "Includes/Enum.stencil" %}
