--- Templates/Swift/Sources/Model.swift
@@ -1,5 +1,0 @@
-{% include "Includes/Header.stencil" %}
-
-import Foundation
-
-{% include "Includes/Model.stencil" %}
