--- Templates/Swift/template.yml
@@ -1,55 +1,0 @@
-formatter: swift
-options:
-  name: POIAPI # the name of the API
-  fixedWidthIntegers: false # whether to use types like Int32 and Int64
-  mutableModels: true # whether model properties are mutable
-  safeOptionalDecoding: false # set invalid optionals to nil instead of throwing
-  safeArrayDecoding: false # filter out invalid items in array instead of throwing
-  modelPrefix: "PC" # applied to model classes and enums
-  modelSuffix: null # applied to model classes
-  modelType: class # can be struct or class
-  modelInheritance: true # must be false for struct modelType
-  modelProtocol: APIModel # the protocol all models conform to
-  modelNames: {} # override model type names
-  enumNames: {} # override enum type names
-  enumUndecodableCase: false # whether to add undecodable case to enums
-  codableResponses: false # constrains all responses/model to be Codable
-  propertyNames: {} # override property names
-  anyType: Any # override Any in generated models
-  numberType: Double # number type without format
-  typeAliases:
-    ID: String
-    DateTime: Date
-    File: Data
-templateFiles:
-  - path: Sources/API.swift
-    destination: "POIAPI.swift"
-  - path: Sources/Coding.swift
-    destination: "Coding.swift"
-  - path: Sources/APIRequest.swift
-    destination: "APIRequest.swift"
-  - path: Sources/APIClient.swift
-    destination: "APIClient.swift"
-  - path: Sources/APIService.swift
-    destination: "APIService.swift"
-  - path: Sources/APIClientError.swift
-    destination: "APIClientError.swift"
-  - path: Sources/APIResponse.swift
-    destination: "APIResponse.swift"
-  - path: Sources/APIResult.swift
-    destination: "APIResult.swift"
-  - path: Sources/RequestBehaviour.swift
-    destination: "RequestBehaviour.swift"
-  - path: Sources/AnyCodable.swift
-    destination: "AnyCodable.swift"
-  - path: Sources/URLEncoding.swift
-    destination: "URLEncoding.swift"
-  - path: Sources/Enum.swift
-    context: enums
-    destination: "Enums/{{ enumName }}.swift"
-  - path: Sources/Model.swift
-    context: schemas
-    destination: "Models/{{ type }}.swift"
-  - path: Sources/Request.swift
-    context: operations
-    destination: "Requests{% if tag %}/{{ tag|upperCamelCase }}{% endif %}/{{ type }}.swift"
