+++ build_specs.sh
@@ -5,5 +5,5 @@ swift build
 for spec in Specs/**
 do
     echo ""
-    sh build_spec.sh ${spec}
+    # sh build_spec.sh ${spec}
 done
